requirements.txt
asgiref==3.8.1
Django==4.2.13
django-crontab==0.7.1
django-filter==24.2
djangorestframework==3.15.1
numpy==1.26.4
pandas==2.2.2
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.1
tzdata==2024.1
